+++ lib/src/components/material_ripple/material_ripple.dart
@@ -2,14 +2,141 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'dart:async';
 import 'dart:html';
+import 'dart:math';
 
 import 'package:angular2/angular2.dart';
 
-import 'src/wave.dart';
-import '../../utils/browser/dom_service/angular_2.dart';
-import '../../utils/disposer/disposer.dart';
+import '../../utils/angular/properties/properties.dart';
+import '../../utils/color/material.dart' show rippleOpacity;
+import '../../utils/browser/feature_detector/feature_detector.dart'
+    show supportsAnimationApi;
+
+// The ripple size is 256x256 for optimal performance (power-of-two texture)
+// while maintaining smooth ripple edges (since the texture is being scaled on
+// the GPU). 256x256 results in 260 KB of GPU memory per ripple.
+const _rippleDiameter = 256;
+const _rippleRadius = _rippleDiameter / 2;
+const _minOpacity = 0.0;
+const _maxOpacity = rippleOpacity;
+
+// After 3 ripples have been created, older ones will be recycled to save GPU
+// memory and DOM memory. This means that there can be a maximum of 3 ripples at
+// a time.
+const _maxRipples = 3;
+
+// All of these variables are outside of the class because it results in faster
+// variable access patterns.
+int _numRipples = 0;
+int _rippleIndex = 0;
+
+// If these were initialized here (and final), dart2js would wait to initialize
+// them until the first mousedown event, increasing latency.
+List<DivElement> _ripplePool;
+DivElement _rippleTemplate;
+Map<String, double> _opacityTiming;
+List<Map<String, double>> _opacityKeyframes;
+Map<String, dynamic> _transformTiming;
+
+// This is outside of the class because it causes dart2js to inline this
+// function and use faster variable access patterns.
+void _createRipple(MouseEvent e, HtmlElement container, bool center) {
+  // All of the DOM reads occur before the DOM writes.
+  final rect = container.getBoundingClientRect();
+  final clientX = e.client.x;
+  final clientY = e.client.y;
+
+  // Create a ripple or grab one from the pool.
+  var ripple;
+  if (_numRipples < _maxRipples) {
+    ripple = _rippleTemplate.clone(false) as DivElement;
+    _ripplePool[_rippleIndex] = ripple;
+    _numRipples++;
+  } else {
+    ripple = _ripplePool[_rippleIndex];
+    ripple.remove();
+  }
+
+  if (++_rippleIndex == _maxRipples) _rippleIndex = 0;
+
+  // Apply the animation and append the ripple to the container.
+  // The applyAnimation functions are inlined by dart2js.
+  if (supportsAnimationApi) {
+    _applyAnimation(ripple, center, rect, clientX, clientY);
+  } else {
+    _applyFallbackAnimation(ripple, center, rect, clientX, clientY);
+  }
+  container.append(ripple);
+}
+
+/// Dynamically generate and apply the ripple animation.
+void _applyAnimation(
+    DivElement ripple, bool center, Rectangle rect, int clientX, int clientY) {
+  // The ripple starting diameter is 60% of the largest container dimension.
+  final containerWidth = rect.width;
+  final containerHeight = rect.height;
+  final maxDimension =
+      (containerWidth > containerHeight) ? containerWidth : containerHeight;
+  final minScale = (maxDimension * 0.6) / _rippleDiameter;
+
+  // The ripple ends 10px larger than the container.
+  // This assumes a rectangular container.
+  final maxRadius =
+      sqrt(pow(containerWidth / 2, 2) + pow(containerHeight / 2, 2)) + 10;
+  final maxScale = maxRadius / _rippleRadius;
+
+  var top;
+  var left;
+  var initialTransform;
+  var finalTransform;
+
+  if (center) {
+    top = 'calc(50% - ${_rippleRadius}px)';
+    left = 'calc(50% - ${_rippleRadius}px)';
+    initialTransform = 'scale($minScale)';
+    finalTransform = 'scale($maxScale)';
+  } else {
+    final offsetX = clientX - rect.left - _rippleRadius;
+    final offsetY = clientY - rect.top - _rippleRadius;
+    // The ripple drifts from the point of touch to the center of the container.
+    final driftX = (containerWidth / 2 - _rippleRadius) - offsetX;
+    final driftY = (containerHeight / 2 - _rippleRadius) - offsetY;
+    top = '${offsetY}px';
+    left = '${offsetX}px';
+    initialTransform = 'translate(0, 0) scale($minScale)';
+    finalTransform = 'translate(${driftX}px, ${driftY}px) scale($maxScale)';
+  }
+
+  final transformKeyframes = [
+    {'transform': initialTransform},
+    {'transform': finalTransform},
+  ];
+
+  ripple.style.cssText = 'top: $top; left: $left; transform: $finalTransform';
+  ripple.animate(_opacityKeyframes, _opacityTiming);
+  ripple.animate(transformKeyframes, _transformTiming);
+}
+
+/// Apply a static fallback animation for browsers that don't support the
+/// Web Animations API.
+void _applyFallbackAnimation(
+    DivElement ripple, bool center, Rectangle rect, int clientX, int clientY) {
+  var top;
+  var left;
+
+  if (center) {
+    top = 'calc(50% - ${_rippleRadius}px)';
+    left = 'calc(50% - ${_rippleRadius}px)';
+  } else {
+    final offsetX = clientX - rect.left - _rippleRadius;
+    final offsetY = clientY - rect.top - _rippleRadius;
+    top = '${offsetY}px';
+    left = '${offsetX}px';
+  }
+
+  ripple.style.top = top;
+  ripple.style.left = left;
+}
 
 /// A component that provides a visual effect that other elements can use to
 /// simulate a rippling effect emanating from the point of contact.
@@ -17,204 +144,73 @@ import '../../utils/disposer/disposer.dart';
 /// The effect can be visualized as a concentric circle with motion.
 ///
 /// Example of use:
-///     <div><material-ripple></material-ripple></div>
+///     <div style="position: relative">
+///       <material-ripple></material-ripple>
+///     </div>
 ///
-/// `material-ripple` listens to `mouse-down` and `mouse-up` events so it would
-/// display ripple effects when touches are on it.
+/// The containing element needs to be position: relative.
 ///
 /// By default, the ripple is centered on the point of contact.
 @Component(
     selector: 'material-ripple',
-    providers: const [domServiceBinding],
-    host: const {'(mousedown)': r'downAction($event)'},
-    template: '', // Removed for performance.
+    template: '',
     encapsulation: ViewEncapsulation.None,
     styleUrls: const ['material_ripple.scss.css'],
     changeDetection: ChangeDetectionStrategy.Detached)
 class MaterialRippleComponent implements OnDestroy {
-  static const String FOCUSED_OPACITY = '.12';
-  static const double FOCUSED_OPACITY_INT = 0.12;
-  static const String TRANSPARENT_OPACITY = '0';
+  final HtmlElement _element;
+  EventListener _onMouseDown;
 
-  HtmlElement _element;
-  final Disposer _disposer = new Disposer.multi();
-  HtmlElement _background;
-  HtmlElement _container;
-  final DomService _domService;
+  MaterialRippleComponent(ElementRef elementRef)
+      : _element = elementRef.nativeElement {
+    // These are initialized here instead of when they're declared because
+    // dart2js would otherwise wait to initialize them until they are used.
+    _ripplePool ??= new List<DivElement>(_maxRipples);
+    _opacityTiming ??= {
+      'duration': 418.0,
+    };
+    _opacityKeyframes ??= [
+      {'opacity': _minOpacity},
+      {'opacity': _maxOpacity, 'offset': 0.2},
+      {'opacity': _maxOpacity, 'offset': 0.4},
+      {'opacity': _minOpacity},
+    ];
+    _transformTiming ??= {
+      'duration': 333.0,
+      'easing': 'cubic-bezier(0.4, 0.0, 0.2, 1)',
+    };
+    // This is className = instead of classes.add because classes.add compiles
+    // to a DOM read, conversion to List, and then a DOM write.
+    if (_rippleTemplate == null) {
+      final className =
+          (supportsAnimationApi) ? '__acx-ripple' : '__acx-ripple fallback';
+      _rippleTemplate = new DivElement()..className = className;
+    }
 
-  final bool _center;
-  final bool _recenter;
-  final _waves = <Wave>[];
-
-  bool _isAnimating = false;
-  StreamController<bool> _onAnimatingController;
-
-  bool _focused = false;
-
-  /// Creates a new material ripple at [elementRef].
-  factory MaterialRippleComponent(
-      @Attribute('center') String center,
-      @Attribute('recenter') String recenter,
-      ElementRef elementRef,
-      DomService domService) {
-    return new MaterialRippleComponent._(
-        elementRef, domService, center != null, recenter != null);
+    // This is necessary because if _onMouseDown was a method, a new closure
+    // would be created each time it was referenced. That means that
+    // _onMouseDown in removeEventListener would point to a different listener,
+    // so the listener would not be removed.
+    _onMouseDown = (e) {
+      // This is inlined by dart2js so we aren't incurring an additional
+      // function call here.
+      _createRipple(e, _element, _center);
+    };
+    // This is about 5x faster than _element.onMouseDown.listen or Angular
+    // (mousedown) because this compiles directly to addEventListener, whereas
+    // the streams approach adds several layers of slow indirection.
+    _element.addEventListener('mousedown', _onMouseDown);
   }
 
-  MaterialRippleComponent._(
-      ElementRef elementRef, this._domService, this._center, this._recenter)
-      : _element = elementRef.nativeElement;
+  /// Whether the ripple should start from the center of the container.
+  bool _center = false;
+  @Input('center')
+  set center(value) {
+    _center = getBool(value);
+  }
 
   @override
   void ngOnDestroy() {
-    _disposer.dispose();
-    _element = null;
-    _background = null;
-    _container = null;
+    _element.removeEventListener('mousedown', _onMouseDown);
   }
-
-  /// Trigger animation.
-  ///
-  /// All of this is one big DOM Write. Ensure that no DOM reads occur in this
-  /// function, and schedule it's execution with a [DomService.scheduleWrite]
-  void _animate() {
-    if (_element == null) return; // Component was destroyed.
-    if (!_isAnimating) {
-      _isAnimating = true;
-      if (_onAnimatingController != null) {
-        _onAnimatingController.add(true);
-      }
-    }
-
-    var shouldKeepAnimating = false;
-
-    for (var i = 0; i < _waves.length; i++) {
-      // Draw every wave.
-      final wave = _waves[i];
-      if (!wave.isAnimationComplete) {
-        shouldKeepAnimating = true;
-      }
-
-      wave.draw();
-
-      // Set the background layer opacity.
-      if (_focused && wave.outerOpacity < FOCUSED_OPACITY_INT) {
-        _background.style.opacity = FOCUSED_OPACITY;
-      } else {
-        _background.style.opacity = wave.outerOpacity.toString();
-      }
-
-      // If the ripple is done animating, remove it.
-      if (wave.isOpacityFullyDecayed && !wave.isRestingAtMaxRadius) {
-        _removeWave(wave);
-      }
-    }
-
-    if (!shouldKeepAnimating && _waves.isEmpty) {
-      _triggerAnimationComplete();
-    } else {
-      _domService.nextFrame.then((_) {
-        _domService.scheduleWrite(_animate);
-      });
-    }
-  }
-
-  /// An event stream that fires when the animation state changes.
-  @Output('animating')
-  Stream<bool> get onAnimating {
-    if (_onAnimatingController == null) {
-      _onAnimatingController = new StreamController<bool>(sync: true);
-    }
-    return _onAnimatingController.stream;
-  }
-
-  /// Triggered when there is a browser event like mouse down.
-  void downAction([MouseEvent e]) {
-    _initDom();
-
-    final wave = _addWave()..downAction(e?.client);
-    _nextMouseUp().then((_) {
-      wave.upAction();
-      _domService.scheduleWrite(_animate);
-    });
-    if (!_isAnimating) {
-      _domService.scheduleWrite(_animate);
-    }
-  }
-
-  /// Completes the next time the mouse is released.
-  Future _nextMouseUp() {
-    final completer = new Completer.sync();
-    final maybeComplete = (UIEvent e) {
-      if (completer.isCompleted) return;
-      completer.complete(e);
-      _disposer.dispose(); // Cleanup the document subscriptions.
-    };
-    _disposer.addStreamSubscription(
-        document.onMouseUp.take(1).listen(maybeComplete));
-    _disposer.addStreamSubscription(
-        document.onDragEnd.take(1).listen(maybeComplete));
-    _disposer.addStreamSubscription(
-        document.onTouchEnd.take(1).listen(maybeComplete));
-    return completer.future;
-  }
-
-  /// Trigger the end of animation.
-  void _triggerAnimationComplete() {
-    _isAnimating = false;
-    if (!_focused) {
-      _background.style.opacity = TRANSPARENT_OPACITY;
-    }
-    if (_onAnimatingController != null) {
-      _onAnimatingController.add(false);
-    }
-  }
-
-  void _initDom() {
-    // Initialize the background and waves divs.
-    if (_element != null && _background == null) {
-      _background = new Element.div()
-        ..classes.add('__material-ripple_background');
-      _container = new Element.div()..classes.add('__material-ripple_waves');
-      _element..append(_background)..append(_container);
-    }
-  }
-
-  Wave _addWave() {
-    final wave = new Wave(_container, center: _center, recenter: _recenter);
-    _waves.add(wave);
-    return wave;
-  }
-
-  /// Remove a completed [wave].
-  void _removeWave(Wave wave) {
-    // Remove the DOM element.
-    wave.remove();
-
-    // Remove from the collection of managed waves.
-    _waves.remove(wave);
-  }
-
-  /// Sets the focus of the ripple.
-  @Input()
-  set focused(bool value) {
-    if (_focused == value) return;
-    _focused = value;
-    _initDom();
-
-    // Defensively check for the element being created before scheduling an
-    // opacity change - if another dom write triggers change detection then a
-    // ripple can be created in a DOM write which in turn will process this
-    // synchronously.
-    if (!_isAnimating && _background != null) {
-      _domService.scheduleWrite(() {
-        if (_background != null) {
-          _background.style.opacity =
-              _focused ? FOCUSED_OPACITY : TRANSPARENT_OPACITY;
-        }
-      });
-    }
-  }
-
-  bool get focused => _focused;
 }