+++ CHANGELOG.md
@@ -1,8 +1,40 @@
+## 0.3.0-alpha
+
+This code is considered production quality, but depends on angular2: 3.0.0-alpha.
+The alpha tag represents the evolving nature of the AngularDart api.
+
+ * Add Material Tooltip.
+ * Material Ripple:
+   * Add GPU acceleration.
+   * `rippleBindings` have been removed as they are no longer used.
+ * Internal updates for compatibility with Angular 3.0.0-alpha.
+ * Material Expansion Panel:
+   * Fix CSS rule that causes header text to turn gray on hover/focus.
+   * Support auto-focus on a content element when the material expansion panel expands.
+   * Fix Yes/No button ordering.
+ * Material Input:
+   * Add a blur update value accessor.
+   * Add multiple attribute.
+   * Remove unused properties: rows and maxRows.
+ * Material Input Multiline: Add auto grow in size.
+ * Material Popup: Update change detection for OnPush.
+ * Material Progress: Update to animate when main thread is blocked.
+ * Material Radio: Adjust size to 24px.
+ * Material Yes/No: Add toggle for yes button visibility.
+ * Scorecard: Update change detection.
+ * Fix flipped alignment positions when isRtl is used.
+ * Fix popup event handling.
+ * Remove 'uninitialized' as a default value.
+ * Remove unused CSS rules.
+ * Update styles to meet Material UI spec.
+ * Bug fixes.
+ * Strong Mode fixes.
+
 ## 0.2.2
 
  * Add Material Popup, a basic popup component.
  * Update Material Checkbox icon size.
- * Cleanup framework stabilizers since [issue #24843](https://github.com/dart-lang/sdk/issues/24843) 
+ * Cleanup framework stabilizers since [issue #24843](https://github.com/dart-lang/sdk/issues/24843)
    in the Dart SDK has been resolved.
  * Remove unused files.
 
@@ -15,7 +47,7 @@
 
 * Add a modal dialog window called material_dialog.
 * Rename MultilineMaterialInputComponent to MaterialMultilineInputComponent.
-* Make the deferredContent placeholder element optional, off by default. 
+* Make the deferredContent placeholder element optional, off by default.
 * Remove InputTextModel.
 * Cleanup linter warnings.
 * Refactor color support.